scraper/requirements.txt
google-generativeai==0.8.3
requests==2.32.3
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.3.0
aiohttp==3.11.11
asyncio==3.4.3
pydantic==2.10.4